requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
requests==2.31.0
huggingface_hub==0.24.6
